requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers
django-getenv
djangorestframework-simplejwt==5.3.0
psycopg2-binary
injector==0.21.0
python-dotenv~=1.1.0
pandas
openpyxl
django-oauth-toolkit==3.0.1
requests~=2.32.3
django-allauth
dj_rest_auth